+++ NEWS.md
@@ -1,6 +1,8 @@
 ## CHANGES IN udpipe VERSION 0.8.6
 
 - Downloading models to paths containing non-ASCII characters now works (issue #95)
+- strsplit.data.frame gains ... which are passed on to strsplit (e.g. to use fixed=TRUE for speeding up)
+- read_connlu is now using fixed=TRUE when splitting by newline symbol (for speeding up)
 
 ## CHANGES IN udpipe VERSION 0.8.5